requirements.txt
google-cloud-bigquery
requests
black
pylint